add wf macro, move macros to separate module
